Updates for new repository structure (#4)
TL;DR
-----

Incorporates MetalLB since it's required for all clusters

Details
-------

Supports the new lab repository structure by pulling in MetalLB
and adding vendoring for the components being installed. In the
earlier structure MetalLB was being installed into the management
cluster, but it's needed in every cluster so I'm moving it here.